// job-diva-frontend/src/types/auth.types.ts
export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  linkedInId?: string;
  linkedInAccessToken?: string;
}

export interface AuthState {
  user: User | null;
  accessToken: string | null;
  isAuthenticated: boolean;
  loading: boolean;
  error: string | null;
}

export interface LoginResponse {
  user: User;
  accessToken: string;
  refreshToken: string;
}

export interface LoginCredentials {
  email: string;
  password: string;
}

export interface RegisterData {
  name: string;
  email: string;
  password: string;
}